Move CodableFeedStore to its own file in production code
